feat(dragdrop): types to file
