+++ src/components/DragDropItem/DragDropItem.tsx
@@ -1,8 +1,8 @@
 import clsx from 'clsx';
 import React from 'react';
 import { Draggable } from 'react-beautiful-dnd';
-import { ItemType } from '../DragDrop/DragDrop';
 import styles from '../DragDrop/DragDrop.module.css';
+import { ItemType } from '../DragDrop/DragDropTypes';
 
 export interface Props {
   /**
